pyproject.toml
[project]
name = "curriculo-facil"
version = "1.0.0"
description = "Gerador de currículos profissionais com interface gráfica"
requires-python = ">=3.10"
dependencies = [
    "fpdf2>=2.7"
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
